Fix average calculation in sum and average array program

+++ Playgrounds/Sum_Avg_Array/main.c
@@ -10,7 +10,7 @@ void main(){
         scanf("%d",&a[i]);
     for(i=0;i<n;i++)
         sum=sum+a[i];
-    avg=sum/n;
+    avg=(float)sum/n;
     printf("\nSUM=%d",sum);
     printf("\nAverage=%f\n",avg);
 }
